requirements.txt
# Mrpbot 2.0 依赖

# 核心
pyyaml>=6.0.1
python-dotenv>=1.0.1

# 网络
websockets>=12.0
httpx>=0.27.0

# AI
openai>=1.12.0

# 异步
asyncio>=3.4.3

# 工具
aiofiles>=23.2.1

# 测试 (可选)
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Web 管理 (可选)
fastapi>=0.109.0
uvicorn>=0.27.0
